refactor: update CalculatorMath precision and enhance tooltip functionality in RecordCard component
- Changed precision in CalculatorMath from 1025 to 64 for improved performance.
- Added ToolTip component to RecordCard for displaying additional information on hover.
- Refactored tooltip handling logic to separate memo and result tooltips for better clarity.
- Improved styling for memo and record text in RecordCard for enhanced readability.

These changes aim to optimize mathematical operations and enhance user interaction through improved tooltip functionality.

+++ src/classes/CalculatorMath.ts
@@ -7,7 +7,7 @@ import { all, create } from 'mathjs';
 // MathJS 라이브러리 설정
 export const MathB = create(all as FactoryFunctionMap, {
   number: 'BigNumber',
-  precision: 1025,
+  precision: 64,
 });
 
 // BigNumber 타입 정의
